Move `outline_config` to `gpu_bridge`
