+++ crates/viewer/re_viewer_context/src/gpu_bridge/mod.rs
@@ -3,6 +3,7 @@
 mod colormap;
 mod image_to_gpu;
 mod instance_hash_conversions;
+mod outline;
 mod re_renderer_callback;
 
 pub use colormap::{colormap_edit_or_view_ui, colormap_to_re_renderer};
@@ -13,6 +14,7 @@ pub use image_to_gpu::{
 pub use instance_hash_conversions::{
     instance_path_hash_from_picking_layer_id, picking_layer_id_from_instance_path_hash,
 };
+pub use outline::outline_config;
 pub use re_renderer_callback::new_renderer_callback;
 
 use crate::TensorStats;
